+++ plc4go/protocols/bacnetip/readwrite/model/BACnetContextTagBoolean.go
@@ -237,7 +237,7 @@ func (m *_BACnetContextTagBoolean) SerializeWithWriteBuffer(ctx context.Context,
 
 		// Simple Field (value)
 		value := uint8(m.GetValue())
-		_valueErr := writeBuffer.WriteUint8("value", 8, (value))
+		_valueErr := writeBuffer.WriteUint8("value", 8, uint8((value)))
 		if _valueErr != nil {
 			return errors.Wrap(_valueErr, "Error serializing 'value' field")
 		}
