+++ plc4go/protocols/bacnetip/readwrite/model/BVLCForwardedNPDU.go
@@ -253,7 +253,7 @@ func (m *_BVLCForwardedNPDU) SerializeWithWriteBuffer(ctx context.Context, write
 		}
 		for _curItem, _element := range m.GetIp() {
 			_ = _curItem
-			_elementErr := writeBuffer.WriteUint8("", 8, _element)
+			_elementErr := writeBuffer.WriteUint8("", 8, uint8(_element))
 			if _elementErr != nil {
 				return errors.Wrap(_elementErr, "Error serializing 'ip' field")
 			}
@@ -264,7 +264,7 @@ func (m *_BVLCForwardedNPDU) SerializeWithWriteBuffer(ctx context.Context, write
 
 		// Simple Field (port)
 		port := uint16(m.GetPort())
-		_portErr := writeBuffer.WriteUint16("port", 16, (port))
+		_portErr := writeBuffer.WriteUint16("port", 16, uint16((port)))
 		if _portErr != nil {
 			return errors.Wrap(_portErr, "Error serializing 'port' field")
 		}
